+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 * Add stage export.
 
+### Fixed
+* Unlock empty object. When object creating.
+
 ## [0.14.0-beta02]
 ### Fixed
 * Fix locales