+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 * Add stage export.
+### Added
+* Add panel of class dependencies diagrams on class sheet.
 
 ### Fixed
 * Unlock empty object. When object creating.